donation_manager/donation_app/requirements.txt
Django==4.2
Pillow==10.0.0
django_crispy_forms==2.0
django_filter==23.2